pyproject.toml
# [build-system]
# requires = ["flit_core >=3.2,<4"]
# build-backend = "flit_core.buildapi"
[build-system]
requires = ["hatchling >= 1.29", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[project]
name = "cpa"
version = "0.0.1"
description = "Copy-Paste Augmentation for ADLCV 2026 Spring"
authors = [
  { name = "Alireza Hajebrahimi", email = "alihaj@dtu.dk" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",

]
dependencies = [
    "accelerate>=1.13.0",
    "albumentationsx[headless]>=2.1.3",
    "datasets>=4.8.4",
    "diffusers>=0.37.1",
    "einops>=0.8.2",
    "faster-coco-eval>=1.7.0",
    "hatchling>=1.29.0",
    "hydra-core>=1.3.2",
    "ipython",
    "jupyterlab",
    "lightning>=2.6.1",
    "loguru",
    "matplotlib",
    "notebook",
    "numpy",
    "pandas",
    "pillow>=12.2.0",
    "pip",
    "pycocotools>=2.0.11",
    "pytest",
    "python-dotenv",
    "ruff",
    "scikit-image>=0.26.0",
    "scikit-learn",
    "torch",
    "torchdiffeq>=0.2.5",
    "torchvision",
    "tqdm",
    "transformers>=5.5.4",
    "typer",
    "ultralytics>=8.4.41",
    "wandb>=0.26.0",
]
requires-python = "~=3.13.0"


[tool.ruff]
line-length = 110
src = ["cpa"]
include = ["pyproject.toml", "cpa/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["cpa"]
force-sort-within-sections = true
